removed a bunch of %d (int) vs magma_int_t warnings
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@4803 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ sparse-iter/control/magma_z_vio.cpp
@@ -56,7 +56,7 @@ magma_z_vvisu(      magma_z_vector x,
                     magma_int_t displaylength ){
 
     printf("visualize entries %d - %d of vector ", 
-                    offset, offset+displaylength );
+                    (int) offset, (int) (offset+displaylength) );
     fflush(stdout);  
     if( x.memory_location == Magma_CPU ){
         printf("located on CPU:\n");
